+++ acsylla/_cython/statement/prepared.pyx
@@ -17,7 +17,7 @@ cdef class PreparedStatement:
         prepared.serial_consistency = serial_consistency
         return prepared
 
-    def bind(self, object page_size=None, object page_state=None, timeout=None, consistency=None, serial_consistency=None):
+    def bind(self, object page_size=None, object page_state=None, timeout=None, consistency=None, serial_consistency=None, execution_profile=None):
         cdef CassStatement* cass_statement
         cdef Statement statement
 
@@ -30,5 +30,14 @@ cdef class PreparedStatement:
             timeout or self.timeout,
             consistency or self.consistency,
             serial_consistency or self.serial_consistency,
+            execution_profile
         )
+        if execution_profile is not None:
+            self.set_execution_profile(cass_statement, execution_profile)
         return statement
+
+    cdef set_execution_profile(self, CassStatement* cass_statement, str name):
+        if name is None:
+            return
+        cdef CassError error = cass_statement_set_execution_profile(cass_statement, name.encode())
+        raise_if_error(error)