+++ CHANGELOG.md
@@ -6,6 +6,7 @@ NOT RELEASED YET
 - Add metadata support
 - Add logging
 - Add Blacklist, whitelist DC, and blacklist, whitelist hosts load balancing policies
+- Add execution profiles
 - Update cpp-driver to latest version
 - Static link OpenSSL and libuv for MacOS wheel